Break out scrollpos into own file.
